+++ src/LinInter.cpp
@@ -14,13 +14,14 @@ double LinInterp_cpp(NumericVector x, NumericVector y, double xlev) {
 
   NumericVector x2 = pow(x-xlev,2);
   int close = vecminInd(x2);
+  
   double xsb = x[close];
   int close2 = (xsb < xlev)*2-1;
   NumericVector ind(2);
   ind[0] = close;
   ind[1] = close+close2;
   int len = x.length();
-  if (ind[1]>len) {
+  if (ind[1]>len-1) {
     ind[1] = ind[0];
     ind[0] = ind[0]-1;
   }