+++ NEWS.md
@@ -3,6 +3,7 @@ The current version of the `MSEtool` package is available for download from [CRA
 # MSEtool 3.7.3 - Development
 
 ## Fixes
+- fix to `addMPs`
 - fix print for `PM` objects when `nsim<15`
 - fix to `LinInterp_cpp` when searching for values on bounds