+++ NEWS.md
@@ -1,6 +1,6 @@
 The current version of the `MSEtool` package is available for download from [CRAN](https://CRAN.R-project.org/package=MSEtool).
 
-# MSEtool 3.7.3 - Development
+# MSEtool 3.7.3
 
 ## Fixes
 - fix to `addMPs`
@@ -8,7 +8,7 @@ The current version of the `MSEtool` package is available for download from [CRA
 - fix to `LinInterp_cpp` when searching for values on bounds
 
 ## New additions
-- add `SAM2OM`, `Splot`, and empirical MP `Emp`
+- add `Splot` and empirical MP `Emp`
 
 # MSEtool 3.7.2